feat: implement gender selection screen and navigation
Implemented a Gender Selection screen with male and female options represented by icons. The screen includes a "Continue" button that is enabled only after a gender is selected.

Added navigation from the login screen to the Gender Selection screen using Jetpack Compose navigation. Clicking the "Sign In" button on the login screen navigates to the Gender Selection screen. Added a back button on Gender Selection screen to navigate back to Login Screen.